LOGBROKER-10206 Better
