+++ ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/topic/client.h
@@ -49,9 +49,6 @@ public:
     //! Create write session.
     std::shared_ptr<ISimpleBlockingWriteSession> CreateSimpleBlockingWriteSession(const TWriteSessionSettings& settings);
 
-    //! Create simple blocking keyed write session. Experimental feature. DO NOT USE IN PRODUCTION.
-    std::shared_ptr<ISimpleBlockingKeyedWriteSession> CreateSimpleBlockingKeyedWriteSession(const TProducerSettings& settings);
-
     //! Create keyed write session. Experimental feature. DO NOT USE IN PRODUCTION.
     std::shared_ptr<IKeyedWriteSession> CreateKeyedWriteSession(const TProducerSettings& settings);
