+++ ydb/public/sdk/cpp/src/client/topic/impl/topic_impl.h
@@ -316,7 +316,6 @@ public:
     // Runtime API.
     std::shared_ptr<IReadSession> CreateReadSession(const TReadSessionSettings& settings);
     std::shared_ptr<ISimpleBlockingWriteSession> CreateSimpleWriteSession(const TWriteSessionSettings& settings);
-    std::shared_ptr<ISimpleBlockingKeyedWriteSession> CreateSimpleKeyedWriteSession(const TProducerSettings& settings);
     std::shared_ptr<IKeyedWriteSession> CreateKeyedWriteSession(const TProducerSettings& settings);
     std::shared_ptr<IProducer> CreateProducer(const TProducerSettings& settings);
     std::shared_ptr<IWriteSession> CreateWriteSession(const TWriteSessionSettings& settings);
